Added an animation to fade the gray background after 1.5 seconds
Apparently the TapGestureRecognizer for iOS times out if the user holds the tap for an estimated more than 1.5 seconds. When that happens, none of the TouchesEnded, TouchesMoved, TouchesCancelled events are triggered for us to fade the gray background out. So this animation chain will ensure that the background will be faded out.